+++ src/module/data/effect/base.mjs
@@ -1,4 +1,4 @@
-import enrichHTML from "../../utils/enrichHTML.mjs";
+import enrichHTML from "../../utils/enrich-html.mjs";
 import FormulaField from "../fields/formula-field.mjs";
 
 /**
@@ -65,7 +65,7 @@ export default class BaseEffectModel extends foundry.abstract.TypeDataModel {
   /** @inheritdoc */
   async toEmbed(config, options = {}) {
 
-    const enriched = await enrichHTML(this.description.value, { ...options, relativeTo: this.parent });
+    const enriched = await enrichHTML(this.parent.description, { ...options, relativeTo: this.parent });
 
     const embed = document.createElement("div");
     embed.classList.add("draw-steel", this.parent.type);