+++ src/module/data/item/ability.mjs
@@ -3,7 +3,7 @@ import { DrawSteelActiveEffect, DrawSteelActor, DrawSteelChatMessage } from "../
 import { DamageRoll, PowerRoll } from "../../rolls/_module.mjs";
 import FormulaField from "../fields/formula-field.mjs";
 import { setOptions } from "../helpers.mjs";
-import enrichHTML from "../../utils/enrichHTML.mjs";
+import enrichHTML from "../../utils/enrich-html.mjs";
 import DamagePowerRollEffect from "../pseudo-documents/power-roll-effects/damage-effect.mjs";
 import BaseItemModel from "./base.mjs";
 
@@ -95,7 +95,7 @@ export default class AbilityModel extends BaseItemModel {
 
     this.power.characteristic = {
       key: "",
-      value: null,
+      value: -5,
     };
   }
 
@@ -437,8 +437,8 @@ export default class AbilityModel extends BaseItemModel {
         modifiers: options.modifiers,
         targets: [...game.user.targets].reduce((accumulator, target) => {
           accumulator.push({
-            uuid: target.actor.uuid,
-            modifiers: this.getTargetModifiers(target.actor),
+            uuid: target.actor?.uuid ?? "",
+            modifiers: this.getTargetModifiers(target),
           });
           return accumulator;
         }, []),
@@ -519,7 +519,7 @@ export default class AbilityModel extends BaseItemModel {
 
   /**
    * Get the modifiers based on conditions that apply to ability Power Rolls specific to a target
-   * @param {DrawSteelActor} target A target of the Ability Roll
+   * @param {DrawSteelToken} target A target of the Ability Roll
    * @returns {PowerRollModifiers}
    */
   getTargetModifiers(target) {
@@ -528,18 +528,43 @@ export default class AbilityModel extends BaseItemModel {
       edges: 0,
       bonuses: 0,
     };
+    const targetActor = target.actor;
+    const token = canvas.tokens.controlled[0]?.actor === this.actor ? canvas.tokens.controlled[0] : null;
 
     //TODO: ALL CONDITION CHECKS
 
-    // Frightened condition checks
-    if (DrawSteelActiveEffect.isStatusSource(this.actor, target, "frightened")) modifiers.banes += 1; // Attacking the target frightening the actor
-    if (DrawSteelActiveEffect.isStatusSource(target, this.actor, "frightened")) modifiers.edges += 1; // Attacking the target the actor has frightened
+    // Modifiers requiring just the targeted token to have an actor
+    if (targetActor) {
+      // Frightened condition checks
+      if (DrawSteelActiveEffect.isStatusSource(this.actor, targetActor, "frightened")) modifiers.banes += 1; // Attacking the target frightening the actor
+      if (DrawSteelActiveEffect.isStatusSource(targetActor, this.actor, "frightened")) modifiers.edges += 1; // Attacking the target the actor has frightened
 
-    // Grabbed condition check - targeting a non-source adds a bane
-    if (DrawSteelActiveEffect.isStatusSource(this.actor, target, "grabbed") === false) modifiers.banes += 1;
+      // Grabbed condition check - targeting a non-source adds a bane
+      if (DrawSteelActiveEffect.isStatusSource(this.actor, targetActor, "grabbed") === false) modifiers.banes += 1;
+      // Restrained condition check - targeting restrained gets an edge
+      if (targetActor.statuses.has("restrained")) modifiers.edges += 1;
+    }
 
-    // Restrained condition check - targeting restrained gets an edge
-    if (target.statuses.has("restrained")) modifiers.edges += 1;
+    // Modifiers requiring just a controlled token
+    if (token) {
+      // Flanking checks
+      if (this.keywords.has("melee") && this.keywords.has("strike") && token.isFlanking(target)) modifiers.edges += 1;
+    }
+
+    // Modifiers requiring both a controlled token and the targeted token to have an actor
+    if (token && targetActor) {
+      //Taunted checks - attacking a token other than the taunted source while having LOE to the taunted source gets a double bane
+      if (DrawSteelActiveEffect.isStatusSource(this.actor, targetActor, "taunted") === false) {
+        const tauntedSource = fromUuidSync(this.actor.system.statuses.taunted.sources.first());
+        const activeTokens = tauntedSource?.getActiveTokens?.() ?? [];
+
+        for (const tauntedSourceToken of activeTokens) {
+          if (!token.hasLineOfEffect(tauntedSourceToken)) continue;
+          modifiers.banes += 2;
+          break;
+        }
+      }
+    }
 
     return modifiers;
   }
@@ -557,14 +582,6 @@ export default class AbilityModel extends BaseItemModel {
     const restrictions = this.actor.system.restrictions;
     if (restrictions.type.has(this.type)) return true;
     if (restrictions.dsid.has(this._dsid)) return true;
-
-    // Checking if statuses have restricted this ability based on type or _dsid
-    for (const effect of CONFIG.statusEffects) {
-      if (!this.actor.statuses.has(effect.id) || !effect.restrictions) continue;
-
-      if (effect.restrictions.type?.has(this.type)) return true;
-      if (effect.restrictions.dsid?.has(this._dsid)) return true;
-    }
 
     return false;
   }