+++ src/module/utils/_module.mjs
@@ -1,4 +1,5 @@
 export { default as constructHTMLButton } from "./construct-html-button.mjs";
-export { default as enrichHTML } from "./enrichHTML.mjs";
+export { default as enrichHTML } from "./enrich-html.mjs";
+export { default as evaluateFormula } from "./evaluate-formula.mjs";
 export { default as ModelCollection } from "./model-collection.mjs";
 export { default as selectedActors } from "./selected-actors.mjs";